Standardize Wi-Fi Network Scans using the Bridge Router
Add the `archer-c5-v4/common` module and the `scanForWifiNetwork()` helper function,
which will be used to standardize the process of scanning for the Main Router Wi-Fi Network
using the Bridge Router.

While this commit begins the process of migrating the `puppeteer` and `cgi` methods
over to the `scanForWifiNetwork()` function, the migration will not be completed until
both methods have been sufficiently tested.