Add `ConfigurationOptionsType` & `ProgramVariablesType`
Add the `ConfigurationOptionsType` and `ProgramVariablesType` types
to the `ProgramConfiguration` namespace, which are to be used in place
of `ConfigurationOptions` and `ProgramVariables` when generic arguments
are being used.

These types also help avoid "infinite type instantiation" errors when
the generic object type is passed to other types such as
`ObjectUtils.ComplexObjectKey`.